Renamed detection namescape to pattern-detection.
