+++ scripts/pattern-detection/menu-setup.mjs
