+++ scripts/pattern-detection/pattern-scanner.mjs
