+++ scripts/pattern-detection/patterns.mjs
