+++ templates/pattern-detection/pattern-prompt.hbs
